Move text related serialization logic to a dedicated file.
