+++ src/Fixie.Assertions/CsonSerializer.cs
@@ -1,68 +1,8 @@
-﻿using static Fixie.Assertions.StringUtilities;
-using static System.Environment;
-
-namespace Fixie.Assertions;
+﻿namespace Fixie.Assertions;
 
 partial class CsonSerializer
 {
     static string Serialize(bool x) => x ? "true" : "false";
-
-    static string Serialize(char x) => $"'{Escape(x)}'";
-
-    static string Serialize(Guid x) => Serialize(x.ToString());
-
-    static string Serialize(string x)
-    {
-        if (IsMultiline(x))
-        {
-            var terminal = RawStringTerminal(x);
-
-            return $"{terminal}{NewLine}{x}{NewLine}{terminal}";
-        }
-        
-        return $"\"{string.Join("", x.Select(Escape))}\"";
-    }
-
-    static string RawStringTerminal(string x)
-    {
-        var longestDoubleQuoteSequence = 0;
-        var currentDoubleQuoteSequence = 0;
-
-        foreach (var c in x)
-        {
-            if (c != '\"')
-            {
-                currentDoubleQuoteSequence = 0;
-                continue;
-            }
-
-            currentDoubleQuoteSequence++;
-            if (currentDoubleQuoteSequence > longestDoubleQuoteSequence)
-                longestDoubleQuoteSequence = currentDoubleQuoteSequence;
-        }
-
-        return new string('\"', longestDoubleQuoteSequence < 3 ? 3 : longestDoubleQuoteSequence + 1);
-    }
-
-    static string Escape(char x) =>
-        x switch
-        {
-            '\0' => @"\0",
-            '\a' => @"\a",
-            '\b' => @"\b",
-            '\t' => @"\t",
-            '\n' => @"\n",
-            '\v' => @"\v",
-            '\f' => @"\f",
-            '\r' => @"\r",
-            //'\e' => @"\e", TODO: Applicable in C# 13
-            ' ' => " ",
-            '"' => @"\""",
-            '\'' => @"\'",
-            '\\' => @"\\",
-            _ when (char.IsControl(x) || char.IsWhiteSpace(x)) => $"\\u{(int)x:X4}",
-            _ => x.ToString()
-        };
 
     static string Serialize(Type type)
     {
@@ -74,7 +14,7 @@ partial class CsonSerializer
         return $"typeof({TypeName(type)})";
     }
 
-    private static string TypeName(Type x)
+    static string TypeName(Type x)
         => x switch
             {
                 _ when x == typeof(bool) => "bool",
