+++ src/Fixie.Assertions/CsonWriter.cs
@@ -38,20 +38,20 @@ class CsonWriter(StringBuilder output)
         output.Append(content);
     }
 
-    private void WriteIndentation()
+    void WriteIndentation()
     {
         for (int i = 0;  i < indentation; i++)
             output.Append("  ");
     }
 
-    private void Indent()
+    void Indent()
     {
         indentation++;
         if (indentation > 63)
             throw new CsonException("This type could not be serialized because the object graph is too deep.");
     }
 
-    private void Dedent()
+    void Dedent()
     {
         indentation--;
     }
